requirements.txt
fastapi~=0.128.1
pypdf~=6.6.2
python-docx~=1.2.0
pytesseract
pdf2image
jinja2
python-multipart
uvicorn
